requirements.txt
aiohttp==3.5.4
gevent_socketio==0.3.6
